requirements.txt
streamlit==1.30.0
pandas==2.2.3
plotly==5.14.1
numpy==1.26.4
tensorflow-cpu==2.10.0
scikit-learn==1.3.0
joblib==1.3.2
protobuf==3.20.3

